New value for the "hidden" attribute on items:  "episode" means what "playlist" meant before.
"playlist" now means that items are visible when skipping forward and backward, *unless* the "skipPlaylists" argument is passed as true.

When drawing a guide for a full schedule, one will usually want to pass skipPlaylists=true to skipForward(), cause the entire playlist to be described as one entry in the guide.